+++ src/app/js/model-list.js
@@ -47,6 +47,17 @@ var AttrProto = Y.Attribute.prototype,
     EVT_ADD = 'add',
 
     /**
+    Fired when a model is created or updated via the `create()` method, but
+    before the model is actually saved or added to the list. The `add` event
+    will be fired after the model has been saved and added to the list.
+
+    @event create
+    @param {Model} model The model being created/updated.
+    @since 3.5.0
+    **/
+    EVT_CREATE = 'create',
+
+    /**
     Fired when an error occurs, such as when an attempt is made to add a
     duplicate model to the list, or when a sync layer response can't be parsed.
 
@@ -218,8 +229,7 @@ Y.ModelList = Y.extend(ModelList, Y.Base, {
       instance or a hash of model attributes, in which case a new model instance
       will be created from the hash.
     @param {Object} [options] Options to be passed to the model's `sync()` and
-        `set()` methods and mixed into the `add` event when the model is added
-        to the list.
+        `set()` methods and mixed into the `create` and `add` event facades.
       @param {Boolean} [options.silent=false] If `true`, no `add` event(s) will
           be fired.
     @param {Function} [callback] Called when the sync operation finishes.
@@ -238,9 +248,15 @@ Y.ModelList = Y.extend(ModelList, Y.Base, {
             options  = {};
         }
 
+        options || (options = {});
+
         if (!(model instanceof Y.Model)) {
             model = new this.model(model);
         }
+
+        self.fire(EVT_CREATE, Y.merge(options, {
+            model: model
+        }));
 
         return model.save(options, function (err) {
             if (!err) {